Cleaned up the code, removed the temporary JSON Schema as a better will be constructed later, and made errors specially for both c++ and dart code generation
